backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
django-background-tasks==1.2.8
sqlparse==0.5.3
python-dotenv
requests==2.32.3
